Improve environment variable collection from worker
We are making the environment variable collection less strict in the
sense that we will start collecting all variables that comes from
rhc-worker-script. Meaning that we will trust that what will be passed
down to us is trusted and verified.

This will allow convert2rhel to have access to variables that are set in
rhcd process and are sent down the way to rhc-worker-script.