Fix the get_rollback_failures and JSON gathering bug
In the get_rollback_failures wasn't expected that the ending message of
the rollback section can be present earlier. Fix the bug by checking if
the end message is after the start message of the rollback section.

The JSON gathering bug was caused when convert2rhel exits with 1 and the
JSON wasn't gathered.