Add support for input parameters
This patch introduces the support for input parameters that will come
from the Insights UI. Regarding the environment variables value that
will be passed down to us, there is no change that we need to do in
order to receive it. The only change that was necessary was to adapt the
code to receive the CLI switches. In this case, we are handling it
pretty simple, only identifying the two possibilities for now and
working with it. In the future, if more options appear, we can improve
the code.