Skip environmnet variables that have value as 0
If an environment variable has value as 0, we will skip setting them in
the subprocess. The reason for this is that convert2rhel does not handle
well the case for values in those env vars. Convert2rhel currently only
checks for the presence of the env var in the os.environ object. If it
is set, then it will skip the check, otherwise, it will follow the
execution normally.