Add support for input parameters (#102)
* Add support for input parameters

This patch introduces the support for input parameters that will come
from the Insights UI. Regarding the environment variables value that
will be passed down to us, there is no change that we need to do in
order to receive it. The only change that was necessary was to adapt the
code to receive the CLI switches. In this case, we are handling it
pretty simple, only identifying the two possibilities for now and
working with it. In the future, if more options appear, we can improve
the code.

* Skip environmnet variables that have value as 0

If an environment variable has value as 0, we will skip setting them in
the subprocess. The reason for this is that convert2rhel does not handle
well the case for values in those env vars. Convert2rhel currently only
checks for the presence of the env var in the os.environ object. If it
is set, then it will skip the check, otherwise, it will follow the
execution normally.